refactor: token utils (#1624)
* remove is index token

* refactor: move is native currency

* refactor: is token pair tradable

* refactor: move available functions

* refactor: remove get token address

* Revert "refactor: remove get token address"

This reverts commit 8053add3d7b85fd94683ab33bcf8e9af495a7398.

* refactor: get token address to use tokenlists
