+++ src/components/swap/index.tsx
@@ -44,13 +44,20 @@ export const Swap = (props: SwapProps) => {
   ])
   const { logEvent } = useAnalytics()
   const requiresProtection = useProtection()
-  const isTradablePair = useMemo(
-    () => isTokenPairTradable(requiresProtection, inputToken, outputToken),
-    [requiresProtection, inputToken, outputToken],
-  )
   const { chainId } = useNetwork()
   const { slippage } = useSlippage()
   const { address } = useWallet()
+
+  const isTradablePair = useMemo(
+    () =>
+      isTokenPairTradable(
+        requiresProtection,
+        inputToken.symbol,
+        outputToken.symbol,
+        chainId ?? 1,
+      ),
+    [chainId, requiresProtection, inputToken, outputToken],
+  )
 
   const {
     isOpen: isBuyModalOpen,
